ai-services/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.1

# AI/ML Libraries
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8
pdf2image==1.16.3

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
speech-recognition==3.10.0
gtts==2.4.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Text Processing
textblob==0.17.1
polyglot==16.7.4
googletrans==4.0.0rc1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
celery==5.3.4
kombu==5.3.4
flower==2.0.1
schedule==1.2.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Turkish Language Specific
zeyrek==0.1.2
turkish-stemmer==1.0.2

# Additional ML Tools
huggingface-hub==0.19.4
datasets==2.15.0
evaluate==0.4.1
accelerate==0.25.0
